+++ src/Passwords/Argon2PasswordHasher.php
@@ -4,6 +4,7 @@ namespace Orisai\Auth\Passwords;
 
 use Orisai\Utils\Dependencies\Dependencies;
 use Orisai\Utils\Dependencies\Exception\ExtensionRequired;
+use SensitiveParameter;
 use function assert;
 use function password_hash;
 use function password_needs_rehash;
@@ -42,7 +43,11 @@ final class Argon2PasswordHasher implements PasswordHasher
 		$this->threads = $threads ?? 4;
 	}
 
-	public function hash(string $raw): string
+	// phpcs:ignore SlevomatCodingStandard.Classes.RequireSingleLineMethodSignature
+	public function hash(
+		#[SensitiveParameter]
+		string $raw
+	): string
 	{
 		$hash = password_hash($raw, PASSWORD_ARGON2ID, $this->getOptions());
 		assert($hash !== false); // Since php 7.4 password_hash cannot return false
@@ -60,7 +65,11 @@ final class Argon2PasswordHasher implements PasswordHasher
 		return password_needs_rehash($hashed, PASSWORD_ARGON2ID, $this->getOptions());
 	}
 
-	public function isValid(string $raw, string $hashed): bool
+	public function isValid(
+		#[SensitiveParameter]
+		string $raw,
+		string $hashed
+	): bool
 	{
 		if (!$this->isArgonHashed($hashed)) {
 			return false;